+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Support for `innmind/operating:~5.0`
+
 ### Changed
 
 - Delete calls via the filesystem adapter are done concurrently when possible